DIA oracle support, pod cooldown, fmt

+++ contracts/interfaces/ICamelotRouter.sol
@@ -2,14 +2,14 @@
 pragma solidity >=0.6.2;
 
 interface ICamelotRouter {
-  function factory() external view returns (address);
+    function factory() external view returns (address);
 
-  function swapExactTokensForTokensSupportingFeeOnTransferTokens(
-    uint amountIn,
-    uint amountOutMin,
-    address[] calldata path,
-    address to,
-    address referrer,
-    uint deadline
-  ) external;
+    function swapExactTokensForTokensSupportingFeeOnTransferTokens(
+        uint256 amountIn,
+        uint256 amountOutMin,
+        address[] calldata path,
+        address to,
+        address referrer,
+        uint256 deadline
+    ) external;
 }
